+++ lwjgl/src/main/kotlin/net/yakclient/graphics/lwjgl/event/GL3MouseMoveDispatcher.kt
@@ -2,6 +2,7 @@ package net.yakclient.graphics.lwjgl.event
 
 import net.yakclient.graphics.api.event.MouseMoveData
 import net.yakclient.graphics.api.event.MouseMoveDispatcher
+import net.yakclient.graphics.util.unit.px
 import net.yakclient.graphics.util.use
 import org.lwjgl.glfw.GLFW
 import org.lwjgl.system.MemoryStack.stackPush
@@ -25,7 +26,7 @@ public class GL3MouseMoveDispatcher : MouseMoveDispatcher() {
             val x = xd.toInt()
             val y = yd.toInt()
 
-            val e = MouseMoveData(x, y, x - lastX, y - lastY)
+            val e = MouseMoveData(x.toFloat(), y.toFloat(), (x - lastX).toFloat(), (y - lastY).toFloat())
 
             lastX = x
             lastY = y
